first testing routine for new module setup; added band fill value; fixed wrong layer selection during writing
